chore: preserve user-authored plotting baseline

+++ src/ui/overlay/data_interaction.hpp
@@ -252,12 +252,13 @@ class DataInteraction
 
     // Cached state for drawing
     CursorReadout last_cursor_;
-    Figure*       last_figure_ = nullptr;
-    Axes*         active_axes_  = nullptr;
+    Figure*       last_figure_   = nullptr;
+    Axes*         active_axes_   = nullptr;
     const Axes3D* active_axes3d_ = nullptr;
     Rect          active_viewport_;
-    float         xlim_min_ = 0.0f, xlim_max_ = 1.0f;
-    float         ylim_min_ = 0.0f, ylim_max_ = 1.0f;
+    // Double precision for x: limits can sit at epoch scale (~1.7e9).
+    double xlim_min_ = 0.0, xlim_max_ = 1.0;
+    float  ylim_min_ = 0.0f, ylim_max_ = 1.0f;
 
     // Region selection: remember which axes the ROI was started in
     // so it stays in the correct subplot when the cursor moves.
