+++ src/ui/overlay/tooltip.hpp
@@ -24,7 +24,9 @@ struct NearestPointResult
     bool          found       = false;
     const Series* series      = nullptr;
     size_t        point_index = 0;
-    float         data_x      = 0.0f;
+    // Absolute data x (includes series x_offset); double so epoch-scale
+    // timestamps keep sub-second precision.
+    double        data_x      = 0.0;
     float         data_y      = 0.0f;
     float         data_z      = 0.0f;
     float         screen_x    = 0.0f;
@@ -34,7 +36,7 @@ struct NearestPointResult
     float         dy_dx       = 0.0f;    // Finite-difference derivative at the point
     bool          dy_dx_valid = false;   // True when derivative could be computed
     bool          is_3d       = false;
-    const Axes3D* axes3d     = nullptr;   // Owning 3D axes (for axis labels)
+    const Axes3D* axes3d      = nullptr;   // Owning 3D axes (for axis labels)
 };
 
 // Rich hover tooltip rendered via ImGui over the plot canvas.
@@ -67,12 +69,12 @@ class Tooltip
     void set_theme_manager(ui::ThemeManager* tm) { theme_mgr_ = tm; }
 
    private:
-    ui::ThemeManager* theme_mgr_      = nullptr;
-    ImFont*           font_body_      = nullptr;
-    ImFont*           font_heading_   = nullptr;
-    float             snap_radius_px_   = 8.0f;
+    ui::ThemeManager* theme_mgr_         = nullptr;
+    ImFont*           font_body_         = nullptr;
+    ImFont*           font_heading_      = nullptr;
+    float             snap_radius_px_    = 8.0f;
     float             snap_radius_3d_px_ = 16.0f;
-    bool              enabled_        = true;
+    bool              enabled_           = true;
 
     void draw_3d(const NearestPointResult& nearest,
                  float                     window_width,
